+++ app/src/main/java/com/serhiiromanchuk/utilitybills/data/repositoryimpl/BillRepositoryImpl.kt
@@ -16,7 +16,7 @@ class BillRepositoryImpl @Inject constructor(
         billDao.insertBillItem(mapper.mapEntityToDbModel(billItem))
     }
 
-    override suspend fun deleteBillItem(billItemId: Int) {
+    override suspend fun deleteBillItem(billItemId: Long) {
         billDao.deleteBillItem(billItemId)
     }
